Finished solution

+++ services/doctor-service/src/main/java/be/ucll/da/doctorservice/domain/Doctor.java
@@ -1,3 +1,3 @@
 package be.ucll.da.doctorservice.domain;
 
-public record Doctor(Long id, String fieldOfExpertise, String firstName, String lastName, Integer age, String address) {}
+public record Doctor(Integer id, String fieldOfExpertise, String firstName, String lastName, Integer age, String address) {}
