+++ services/patient-service/src/main/java/be/ucll/da/patientservice/domain/Patient.java
@@ -1,3 +1,3 @@
 package be.ucll.da.patientservice.domain;
 
-public record Patient(Long id, String firstName, String lastName, String email, Boolean isClient) {}
+public record Patient(Integer id, String firstName, String lastName, String email, Boolean isClient) {}
